add comments for some functions and move all polynomial optimizaton related functions to TrixiConvexECOSExt.jl
