+++ src/Trixi.jl
@@ -291,8 +291,8 @@ export PlotData1D, PlotData2D, ScalarPlotData2D, getmesh, adapt_to_mesh_level!,
        adapt_to_mesh_level,
        iplot, iplot!
 
-# PERK-related exports
-export stability_polynomials, bisection
+# Paired Explicit RK-related exports
+export filter_eig_vals, undo_normalization!, stability_polynomials, bisect_stability_polynomial
 
 function __init__()
     init_mpi()
